+++ lib/screens/templates/components/canvas/draggable_image_chip.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math;
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:nahpu/screens/templates/components/canvas/draggable_chip.dart';
+import 'package:nahpu/services/templates/template_canvas_snap_service.dart';
 import 'package:nahpu/screens/templates/template_editor_math.dart';
 import 'package:nahpu/screens/templates/template_model.dart';
