+++ lib/screens/templates/components/canvas/template_canvas_editor.dart
@@ -11,6 +11,7 @@ import 'package:nahpu/screens/templates/components/canvas/draggable_image_chip.d
 import 'package:nahpu/screens/templates/components/canvas/draggable_line_chip.dart';
 import 'package:nahpu/screens/templates/components/canvas/draggable_shape_chip.dart';
 import 'package:nahpu/screens/shared/media/qr.dart' show QrImageView;
+import 'package:nahpu/services/templates/template_canvas_snap_service.dart';
 
 import 'dart:math' as math;
 import 'package:nahpu/screens/templates/template_canvas_stack.dart';
